backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
django-neomodel==0.0.7
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
neo4j==4.4.0
neo4j-driver==4.3.6
neobolt==1.7.17
#neomodel
neomodel==4.0.7
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
stripe==2.63.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
whitenoise==5.3.0
